% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_tennis.R
\name{gg_fulltenniscourt}
\alias{gg_fulltenniscourt}
\title{This function draws a full tennis court.}
\usage{
gg_fulltenniscourt(
  fill_colour = "#3C638E",
  outer_fill = "#ad5049",
  line_colour = "white"
)
}
\value{
A ggplot of a tennis court (dimensions in metres)
}
\description{
This function draws a full tennis court.
}
\examples{
ggplot() + gg_fulltenniscourt()
}
