% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_tennis.R
\name{gg_halftenniscourt}
\alias{gg_halftenniscourt}
\title{This function draws half of a symmetric tennis court.}
\usage{
gg_halftenniscourt(
  fill_colour = "#3C638E",
  outer_fill = "#ad5049",
  line_colour = "white"
)
}
\value{
A ggplot of a tennis court (dimensions in metres)
}
\description{
This function draws half of a symmetric tennis court.
}
\examples{
ggplot() + gg_halftenniscourt()
}
