% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_available_matches}
\alias{load_available_matches}
\title{Load all available matches}
\usage{
load_available_matches(
  data_path = "/Users/petertea/Documents/courtvision_data/data/"
)
}
\arguments{
\item{data_path}{str for saved data URL.}
}
\value{
A Dataframe of a catalogue of available matches (ATP & WTA).
}
\description{
This function loads a dataframe indicating which match_id's are valid for loading data.
}
