% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_available_players}
\alias{load_available_players}
\title{Load all available players}
\usage{
load_available_players(
  data_path = "/Users/petertea/Documents/courtvision_data/data/"
)
}
\arguments{
\item{data_path}{str for saved data URL.}
}
\value{
A Dataframe of a catalogue of available players (ATP & WTA).
}
\description{
This function loads a dataframe indicating which players are valid for loading data.
}
