% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_complete_player_data}
\alias{load_complete_player_data}
\title{For a given player name, load all its PBP and ball trajectory data.}
\usage{
load_complete_player_data(
  player_name,
  data_path = "/Users/petertea/Documents/courtvision_data/data/"
)
}
\arguments{
\item{player_name}{Name of player. Eg: N.DJOKOVIC.}

\item{data_path}{str for saved data URL.}
}
\value{
A Dataframe for all recorded events corresponding to match ID.
}
\description{
This function collects all PBP and ball-trajectory data that a given
player competed.
}
\examples{
load_complete_player_data(player_name = 'F.AUGERALIASSIME')
}
